Adds PredictionWorkflowManager component, the heart of the prediciton application.
This component holds the state for both teams selected, the
map selected, and gamemode selected. These component is comprised
of the other components and utilizes callback functions to return
the selection from its child components. This component drills the
selections to the PredictionOutcomeHandler component which
is then responsible for the api call and displaying predictions.