Refactors the workflow manager component which displays page 1
Data fetching will be triggered in this component to enhance
user experience (when user clicks on make prediciton, there is no longer
a 200 ms delay for fetching the data). This is done via a Memo as well.

Introduces React Navigation for redirecting the user to the prediction page and back.