Revamps page to view predictions (/predictions)
PredictionOutcomeManager -> Now displays predictions for all maps in the user map set selected.
Series winner is at the top, a nav bar can be used to view specific map predictions
MatchOverview -> Row at the top of page displaying the predicted maps won for each team.
PredictionPreview -> window in the middle of the screen showcasing the predictions for a particular map in the match.